third_party/blink/web_tests/fast/js/exceptions-thrown-in-callbacks-expected.txt

CONSOLE ERROR: Uncaught PASS: toString called on exception value thrown from event handler
CONSOLE ERROR: Uncaught PASS: toString called on exception value thrown from sql transaction callback
CONSOLE ERROR: Uncaught PASS: toString called on exception value thrown from sql error callback
CONSOLE ERROR: Uncaught PASS: toString called on exception value thrown from timer
This test ensures that exceptions are handled correctly by the various callback mechanisms present in WebCore.
